ConfigController

+++ backend/src/main/java/es/webapp13/porthub/Repository/PortfolioItemRepository.java
@@ -3,5 +3,5 @@ package es.webapp13.porthub.Repository;
 import es.webapp13.porthub.model.PortfolioItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PortfolioItemRepository extends JpaRepository<PortfolioItem, Integer> {
+public interface PortfolioItemRepository extends JpaRepository<PortfolioItem, Long> {
 }
